+++ crates/re_sdk/src/global.rs
@@ -5,6 +5,6 @@ pub fn global_session() -> std::sync::MutexGuard<'static, Session> {
     use once_cell::sync::OnceCell;
     use std::sync::Mutex;
     static INSTANCE: OnceCell<Mutex<Session>> = OnceCell::new();
-    let mutex = INSTANCE.get_or_init(|| Mutex::new(Session::new()));
+    let mutex = INSTANCE.get_or_init(|| Mutex::new(Session::new(true)));
     mutex.lock().unwrap()
 }